pyproject.toml
[project]
name = "titanic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dash>=3.2.0",
    "dash-html-components>=2.0.0",
    "kaleido>=1.0.0",
    "numpy>=2.3.2",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "plotly>=6.3.0",
]
